Trait tag_game::Agent

pub trait Agent: Sized + Send + Sync {
    type State: Send + Sync;
    type World: World<Self> + Sync;
    fn on_creation(
        &self, 
        id: usize, 
        state: &mut Self::State, 
        world: &Self::World
    ) { ... }

    fn on_update(
        &self, 
        id: usize, 
        state: &mut Self::State, 
        world: &Self::World, 
        population: &[(Self, Self::State)]
    ) { ... }
}

An agent defines, how to act in a simulation.

The Simulation associates the agent with a state, which can be mutated by the agent every tick.

State can be seen as data related to the agent itself. The World is the global state shared between all agents in a simulation.

Associated Types

type State: Send + Sync

The local state associated with the agent

type World: World<Self> + Sync

The global state, provided by the simulation

Provided methods

fn on_creation(&self, id: usize, state: &mut Self::State, world: &Self::World)

Called when an agent is added to the simulation.

The id is a unique id used in the simulation, which is generated when adding the agent to the simulation. The state is the same state passed to Simulation::add_agent().

fn on_update(
&self,
id: usize,
state: &mut Self::State,
world: &Self::World,
population: &[(Self, Self::State)]
)

Called when the simulation is updated.

It retrieves the same parameters as Agent::on_creation() and a list, of all agents, which are currently added to the simulation. population[id] corresponds to this agent.